chore: phase g session start, read v3 evaluation and v4 task description
Phase G — execute task_description_v4.md, 17 work items in 4 stages.
Framing: stop optimizing transcription, start optimizing MV2H emission
and use of free signals (voice IDs already produced by B76, confidences
already produced by every model).

Stage 1 (CPU): G-1..G-7 emitter fixes and published post-processing tricks.
Stage 2 (mostly CPU): G-8..G-12 new signal and diagnostics.
Stage 3 (GPU): G-13 Lakh LoRA, G-14 multi-take, G-15 DDSP solo_flute2.
Stage 4 (close-out): G-16 listening artifact, G-17 Docker harness.

- reports/PHASE_G_PROGRESS.md tracker established (0/17).
- reports/_OOM_INCIDENTS.md placeholder.
- CLAUDE.md and PLAN.md aligned to Phase G framing.
- Carries forward F-1 octave sanity (default-on) and F-2e formant offset
  corrector (opt-in flag) verbatim from Phase E ship.

+++ scripts/exp_B68b_musicgen_lora_audiocraft.py
@@ -86,7 +86,10 @@ def main(n_steps: int = 20, lora_r: int = 16, lr: float = 1e-4,
     print(f"loading audiocraft MusicGen {model_size}")
     t0 = time.time()
     mg = MusicGen.get_pretrained(f"facebook/musicgen-{model_size}", device="cuda")
-    lm = mg.lm  # the autoregressive language model
+    # audiocraft loads in fp16 by default; cast LM to fp32 so LoRA grad math works
+    # without dtype mismatches. Adds ~3 GB VRAM but training stability is worth it.
+    mg.lm = mg.lm.to(torch.float32)
+    lm = mg.lm
     print(f"  loaded in {time.time()-t0:.1f}s; lm params: {sum(p.numel() for p in lm.parameters())/1e6:.1f}M")
 
     # Identify target Linear modules in the LM. audiocraft's LM has cross-attention
